backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
pillow==11.0.0
PyJWT==2.9.0
requests==2.32.3
djoser==2.3.1
setuptools==78.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
drf-extra-fields==3.7.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.8.0rc2
django-templated-mail==1.1.1
filetype==1.2.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
python3-openid==3.2.0
pytz==2025.2
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
gunicorn==23.0.0
tqdm==4.67.1